+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 1.1.0: Maintenance Release
 
 New:
- - _mqtt sparkplug in_ and _mqtt sparkplug device_ node now supports decompression (DEFLATE and GZIP)
+ - _mqtt sparkplug in_ and _mqtt sparkplug device_ node now supports compression (DEFLATE and GZIP)
  - _mqtt sparkplug out_ supports topic defined in input message
  - Invalid mesasges to _mqtt sparkplug out_ without metric types, are not caught and a more friendly error message is now shown.